Fix carousel internal component names.
Clojure's (rings'?) feature to automaticly reload libraries, but without
getting a clean slate first feels like a great way to end up with
impossible state, along with a number of broken commits.

Some might say that the correct way is to always write a bunch of tests.
Lein do encurrage writing tests, going so far to create the test
directories for the user upon project creation. However, having tools
which makes it so easy to err is always a bad idea.

But it's wicked fast when it works!